added tests for nft transfers, fixed logic in transfer fn
