+++ __tests__/nft-keys/nearcon.ava.ts
@@ -1,14 +1,16 @@
 import anyTest, { TestFn } from "ava";
-import { NEAR, NearAccount, tGas, Worker } from "near-workspaces";
+import { KeyPair, NEAR, NearAccount, tGas, Worker } from "near-workspaces";
 import { CONTRACT_METADATA, generateKeyPairs, getDropInformation, getKeyInformation, getKeySupplyForDrop, LARGE_GAS, queryAllViewFunctions, WALLET_GAS } from "../utils/general";
 import { DropConfig, JsonKeyInfo, NFTTokenObject, SimpleData, TokenMetadata } from "../utils/types";
 import { BN } from "bn.js";
-import { readFileSync } from 'fs';
-import * as crypto from "crypto";
+
+import { createDropArgs, generatePerUsePasswords, PasswordPerUse, wrapTxnParamsForTrial } from "./utils";
+
 
 const test = anyTest as TestFn<{
     worker: Worker;
     accounts: Record<string, NearAccount>;
+    keypairs: {keys: KeyPair[], publicKeys: string[]}
   }>;
 
   test.beforeEach(async (t) => {
@@ -39,9 +41,25 @@ const test = anyTest as TestFn<{
         amount: NEAR.parse('10000 N').toString()
     })
 
+    //add 2NEAR to ali's keypom balance
+    await funder.call(keypom, 'add_to_balance', {}, {attachedDeposit: NEAR.parse("5000").toString()});
+        
+    const keys = await generateKeyPairs(50);
+    const basePassword = "MASTER_KEY"
+
+    let passwords: Array<Array<PasswordPerUse>> = await generatePerUsePasswords({
+        publicKeys: keys.publicKeys,
+        basePassword,
+        uses: [1]
+    });
+
+    //create a drop with Ali, doesn't front any cost.
+    await funder.call(keypom, 'create_drop', createDropArgs({pubKeys: keys.publicKeys, passwords, root, wasmDirectory: `${require('path').resolve(__dirname, '..')}/ext-wasm/trial-accounts.wasm`}), {gas: '300000000000000'});
+
     // Save state for test runs
     t.context.worker = worker;
     t.context.accounts = { root, keypom, funder, bob };
+    t.context.keypairs = keys;
 });
 
 // If the environment is reused, use test.after to replace test.afterEach
@@ -51,166 +69,75 @@ test.afterEach(async t => {
     });
 });
 
-// helpers for keypom account contract args
-const PARAM_START = "|kP|";
-const PARAM_STOP = "|kS|";
-export const wrapTxnParamsForTrial = (params: Record<string, string | string[]>) => {
-    let newParams: Record<string, string> = {};
-    // Loop through each key value in params
-    Object.entries(params).forEach(([k, v]) => {
-        // If the value is an array, join it with a comma
-        if (Array.isArray(v)) v = v.join();
-        // Add the key and value to the newParams object
-        newParams[PARAM_START.concat(k)] = v.concat(PARAM_STOP);
-    });
-    return newParams;
-};
+test('Transfer NFT Using Keypom', async t => {
+    const { keypom, funder, bob, root } = t.context.accounts;
+    const keys = t.context.keypairs;
 
-export interface PasswordPerUse {
-    /** The password for this given use */
-    pw: string;
-    /** Which use does the password belong to? These uses are *NOT* zero-indexed so the first use corresponds to `1` not `0`. */
-    key_use: number;
-}
-
-const hashBuf = (str: string, fromHex = false): Promise<ArrayBuffer> =>
-    // return a promise that resolves to a buffer
-    new Promise((resolve, reject) => {
-        // create a hash object
-        const hash = crypto.createHash('sha256');
-        // if the input is hex, convert it to a buffer
-        const buf = fromHex ? Buffer.from(str, 'hex') : Buffer.from(str);
-        // hash the buffer
-        hash.update(buf);
-        // resolve the promise with the hash
-        resolve(hash.digest());
-    }); 
-
-export async function generatePerUsePasswords({
-    publicKeys,
-    uses,
-    basePassword,
-}: {
-    publicKeys: string[];
-    uses: number[];
-    basePassword: string;
-}): Promise<Array<Array<PasswordPerUse>>> {
-    const passwords: Array<Array<PasswordPerUse>> = [];
-
-    // Loop through each pubKey to generate either the passwords
-    for (let i = 0; i < publicKeys.length; i++) {
-        // For each public key, we need to generate a password for each use
-        const passwordsPerUse: Array<PasswordPerUse> = [];
-        for (let j = 0; j < uses.length; j++) {
-            // First inner hash takes in utf8 and returns hash
-            const innerHashBuff = await hashBuf(
-                basePassword + publicKeys[i] + uses[j].toString()
-            );
-            const innerHash = Buffer.from(innerHashBuff).toString('hex');
-
-            // Outer hash takes in hex and returns hex
-            const outerHashBuff = await hashBuf(innerHash, true);
-            const outerHash = Buffer.from(outerHashBuff).toString('hex');
-
-            const jsonPw = {
-                pw: outerHash,
-                key_use: uses[j],
-            };
-            passwordsPerUse.push(jsonPw);
-        }
-        passwords.push(passwordsPerUse);
+    const newKeys = await generateKeyPairs(1);
+    
+    let keyInfo: JsonKeyInfo = await getKeyInformation(keypom, keys.publicKeys[0]);
+    let initialAllowance = keyInfo.allowance;
+    console.log('keyInfo before: ', keyInfo)
+    
+    await keypom.setKey(keys.keys[0]);
+    await keypom.call(keypom, 'nft_transfer', {token_id: `0:0`, receiver_id: funder.accountId, memo: newKeys.publicKeys[0]});
+    try {
+        await keypom.call(keypom, 'nft_transfer', {token_id: `0:0`, receiver_id: bob.accountId, memo: keys.publicKeys[0]});
+        keyInfo = await getKeyInformation(keypom, keys.publicKeys[0]);
+        t.is(1, 2);
+    } catch (e) {
+        t.is(1, 1);
     }
 
-    return passwords;
-}
+    keyInfo = await getKeyInformation(keypom, newKeys.publicKeys[0]);
+    console.log('keyInfo after: ', keyInfo)
+    t.assert(new BN(initialAllowance).gt(new BN(keyInfo.allowance)));
+    t.is(keyInfo.owner_id, funder.accountId);
+});
 
-test('Base Costs', async t => {
+test('Transfer NFT Using Owner Account', async t => {
     const { keypom, funder, bob, root } = t.context.accounts;
+    const keyInfos = t.context.keypairs;
 
-    const actualContracts = Array(10).fill(keypom.accountId);
-    const actualAmounts = Array(10).fill(NEAR.parse("100").toString());
-    const actualMethods = Array(10).fill('create_account_and_claim');
-
-    const attachedDeposit = NEAR.parse("30").toString();
-    const extraAllowance = NEAR.parse("0.1").toString();
-    const basePassword = "MASTER_KEY";
-
-    const {publicKeys} = await generateKeyPairs(50);
-    const wasmDirectory = `${require('path').resolve(__dirname, '..')}/ext-wasm/trial-accounts.wasm`
-    const dropId = Date.now().toString();
-    const createDropArgs = {
-        drop_id: dropId,
-        public_keys: [],
-        deposit_per_use: "0",
-        config: {
-            uses_per_key: 2,
-            nft_key_behaviour: {
-                nft_metadata: {
-                    title: "NEARCON VIP Ticket",
-                    media: 'bafkreib2l2xlbty5uihxgcknqxs5uiinwjx3nif7tk7s32yxcmlr2lbm5i'
-                }
-            }
-        },
-        metadata: JSON.stringify({ticketType: "VIP"}),
-        required_gas: "150000000000000",
-        extra_key_allowance: extraAllowance,
-        fc: {
-            methods: [
-                null,
-                [
-                    {
-                        receiver_id: root.accountId,
-                        method_name: "create_account_advanced",
-                        attached_deposit: attachedDeposit,
-                        args: JSON.stringify({
-                            new_account_id: "INSERT_NEW_ACCOUNT",
-                            options: {
-                                contract_bytes: [...readFileSync(wasmDirectory)],
-                                limited_access_keys: [
-                                    {
-                                        public_key: "INSERT_TRIAL_PUBLIC_KEY",
-                                        allowance: "0",
-                                        receiver_id: "INSERT_NEW_ACCOUNT",
-                                        method_names:
-                                            "execute,create_account_and_claim",
-                                    },
-                                ],
-                            },
-                        }),
-                        user_args_rule: "UserPreferred",
-                    },
-                    {
-                        receiver_id: "",
-                        method_name: "setup",
-                        attached_deposit: "0",
-                        args: JSON.stringify(
-                            wrapTxnParamsForTrial({
-                                contracts: actualContracts,
-                                amounts: actualAmounts,
-                                methods: actualMethods,
-                                funder: "",
-                                repay: "0",
-                                floor: "0",
-                            })
-                        ),
-                        receiver_to_claimer: true,
-                    },
-                ],
-            ],
-        },
-    };
-
-    //add 2NEAR to ali's keypom balance
-    await funder.call(keypom, 'add_to_balance', {}, {attachedDeposit: NEAR.parse("5000").toString()});
+    const newKeys = await generateKeyPairs(1);
     
-    //create a drop with Ali, doesn't front any cost. 
-    await funder.call(keypom, 'create_drop', createDropArgs);
+    await keypom.setKey(keyInfos.keys[0]);
+    await keypom.call(keypom, 'nft_transfer', {receiver_id: funder.accountId, memo: newKeys.publicKeys[0]});
+    
+    let keyInfo: JsonKeyInfo = await getKeyInformation(keypom, newKeys.publicKeys[0]);
+    let initialAllowance = keyInfo.allowance;
+    console.log('keyInfo before: ', keyInfo)
 
-    let passwords: Array<Array<PasswordPerUse>> = await generatePerUsePasswords({
-        publicKeys: publicKeys,
-        basePassword,
-        uses: [1]
-    });
+    await funder.call(keypom, 'nft_transfer', {token_id: `0:0`, receiver_id: bob.accountId, memo: keyInfos.publicKeys[0]});
 
-    await funder.call(keypom, 'add_keys', {drop_id: dropId, public_keys: publicKeys, passwords_per_use: passwords, extra_key_allowance: extraAllowance,}, {gas: '300000000000000'});
+    keyInfo = await getKeyInformation(keypom, keyInfos.publicKeys[0]);
+    console.log('keyInfo after: ', keyInfo)
+    t.assert(new BN(initialAllowance).eq(new BN(keyInfo.allowance)));
+    t.is(keyInfo.owner_id, bob.accountId);
+});
+
+test('Transfer NFT Not Owned By Account', async t => {
+    const { keypom, funder, bob, root } = t.context.accounts;
+    const keyInfos = t.context.keypairs;
+
+    const newKeys = await generateKeyPairs(1);
+
+    await keypom.setKey(keyInfos.keys[0]);
+    await keypom.call(keypom, 'nft_transfer', {receiver_id: funder.accountId, memo: newKeys.publicKeys[0]});
+    
+    let keyInfo: JsonKeyInfo = await getKeyInformation(keypom, newKeys.publicKeys[0]);
+    let initialAllowance = keyInfo.allowance;
+    console.log('keyInfo before: ', keyInfo)
+
+    try {
+        await bob.call(keypom, 'nft_transfer', {token_id: `0:0`, receiver_id: bob.accountId, memo: keyInfos.publicKeys[0]});
+        t.is(1, 2);
+    } catch (e) {
+        t.is(1, 1);
+    }
+
+    keyInfo = await getKeyInformation(keypom, newKeys.publicKeys[0]);
+    console.log('keyInfo after: ', keyInfo)
+    t.assert(new BN(initialAllowance).eq(new BN(keyInfo.allowance)));
+    t.is(keyInfo.owner_id, funder.accountId);
 });
